Move all permissions methods to separate resource
