+++ CakeMail.RestClient/CakeMail.RestClient.csproj
@@ -107,6 +107,7 @@
     <Compile Include="Resources\Campaigns.cs" />
     <Compile Include="Resources\Clients.cs" />
     <Compile Include="Resources\Countries.cs" />
+    <Compile Include="Resources\Permissions.cs" />
     <Compile Include="Utilities\CakeMailIntegerBooleanConverter.cs" />
     <Compile Include="Utilities\Constants.cs" />
     <Compile Include="Utilities\ExtensionMethods.cs" />
