+++ CakeMail.RestClient/CakeMailRestClient.cs
@@ -69,6 +69,7 @@ namespace CakeMail.RestClient
 		public Campaigns Campaigns { get; private set; }
 		public Clients Clients { get; private set; }
 		public Countries Countries { get; private set; }
+		public Permissions Permissions { get; private set; }
 
 		#endregion
 
@@ -106,57 +107,6 @@ namespace CakeMail.RestClient
 			};
 
 			InitializeResources();
-		}
-
-		#endregion
-
-		#region Methods related to PERMISSIONS
-
-		/// <summary>
-		/// Retrieve the list of permissions for a given user
-		/// </summary>
-		/// <param name="userKey">User Key of the user who initiates the call.</param>
-		/// <param name="userId">ID of the user.</param>
-		/// <param name="clientId">ID of the client.</param>
-		/// <returns>An enumeration of permissions</returns>
-		public IEnumerable<string> GetUserPermissions(string userKey, long userId, long? clientId = null)
-		{
-			var path = "/Permission/GetPermissions/";
-
-			var parameters = new List<KeyValuePair<string, object>>()
-			{
-				new KeyValuePair<string, object>("user_key", userKey),
-				new KeyValuePair<string, object>("user_id", userId)
-			};
-			if (clientId.HasValue) parameters.Add(new KeyValuePair<string, object>("client_id", clientId.Value));
-
-			return ExecuteArrayRequest<string>(path, parameters, "permissions");
-		}
-
-		/// <summary>
-		/// Set the permissions granted to a given user
-		/// </summary>
-		/// <param name="userKey">User Key of the user who initiates the call.</param>
-		/// <param name="userId">ID of the user.</param>
-		/// <param name="permissions">Enumeration of permissions</param>
-		/// <param name="clientId">ID of the client.</param>
-		/// <returns>True if the operation succeeded</returns>
-		public bool SetUserPermissions(string userKey, long userId, IEnumerable<string> permissions, long? clientId = null)
-		{
-			var path = "/Permission/SetPermissions/";
-
-			var parameters = new List<KeyValuePair<string, object>>()
-			{
-				new KeyValuePair<string, object>("user_key", userKey),
-				new KeyValuePair<string, object>("user_id", userId)
-			};
-			foreach (var item in permissions.Select((name, i) => new { Index = i, Name = name }))
-			{
-				parameters.Add(new KeyValuePair<string, object>(string.Format("permission[{0}]", item.Index), item.Name));
-			}
-			if (clientId.HasValue) parameters.Add(new KeyValuePair<string, object>("client_id", clientId.Value));
-
-			return ExecuteRequest<bool>(path, parameters);
 		}
 
 		#endregion
@@ -3193,6 +3143,7 @@ namespace CakeMail.RestClient
 			this.Campaigns = new Campaigns(this);
 			this.Clients = new Clients(this);
 			this.Countries = new Countries(this);
+			this.Permissions = new Permissions(this);
 		}
 
 		private static string GetVersion()
